chore(types,client,server): add prefix Mtv to every Room* websockets related types & split in files
