+++ packages/types/src/index.ts
@@ -24,6 +24,7 @@
 
 export * from './http';
 export * from './mtv';
+export * from './mtv-room-websockets';
 export * from './server-to-temporal';
-// export * from './temporal-to-server';
+export * from './user-websockets';
 export * from './websockets';
