+++ packages/types/src/mtv-room-websockets.ts
@@ -1,0 +1,60 @@
+import { MtvWorkflowState } from './mtv';
+
+export interface ChatMessage {
+    author: string;
+    text: string;
+}
+
+export interface ChatClientToServerNewMessageArgs {
+    message: ChatMessage;
+}
+
+export interface ChatServerToClientLoadMessagesArgs {
+    messages: ChatMessage[];
+}
+
+export interface ChatServerToClientReceivedMessageArgs {
+    message: ChatMessage;
+}
+
+export interface ChatClientToServerEvents {
+    NEW_MESSAGE: (args: ChatClientToServerNewMessageArgs) => void;
+}
+
+export interface ChatServerToClientEvents {
+    LOAD_MESSAGES: (args: ChatServerToClientLoadMessagesArgs) => void;
+    RECEIVED_MESSAGE: (args: ChatServerToClientReceivedMessageArgs) => void;
+}
+
+export interface MtvRoomClientToServerCreate {
+    name: string;
+    initialTracksIDs: string[];
+}
+
+export interface MtvRoomClientToServerJoin {
+    roomID: string;
+}
+
+export interface Track {
+    name: string;
+    artistName: string;
+}
+
+export interface MtvRoomClientToServerEvents {
+    CREATE_ROOM: (args: MtvRoomClientToServerCreate) => void;
+    JOIN_ROOM: (args: MtvRoomClientToServerJoin) => void;
+    ACTION_PLAY: () => void;
+    GET_CONTEXT: () => void;
+    ACTION_PAUSE: () => void;
+    GO_TO_NEXT_TRACK: () => void;
+}
+
+export interface MtvRoomServerToClientEvents {
+    CREATE_ROOM_SYNCHED_CALLBACK: (state: MtvWorkflowState) => void;
+    RETRIEVE_CONTEXT: (state: MtvWorkflowState) => void;
+    ACTION_PLAY_CALLBACK: (state: MtvWorkflowState) => void;
+    ACTION_PAUSE_CALLBACK: () => void;
+    CREATE_ROOM_CALLBACK: (state: MtvWorkflowState) => void;
+    FORCED_DISCONNECTION: () => void;
+    JOIN_ROOM_CALLBACK: (state: MtvWorkflowState) => void;
+}
